Make player component. (Process the input.)

+++ Project/CheckMate/Point.cpp
@@ -1,25 +1,25 @@
 #include "Point.h"
 
 namespace Utility {
-	Point::Point(const int x, const int y)
+	Point::Point(const float x, const float y)
 		: m_x(x), m_y(y) {}
 
 	Point::Point(const Point& other)
 		: m_x(other.GetX()), m_y(other.GetY()) {}
 
-	int Point::GetX() const noexcept {
+	float Point::GetX() const noexcept {
 		return m_x;
 	}
 
-	void Point::SetX(int x) noexcept {
+	void Point::SetX(float x) noexcept {
 		m_x = x;
 	}
 
-	int Point::GetY() const noexcept {
+	float Point::GetY() const noexcept {
 		return m_y;
 	}
 
-	void Point::SetY(int y) noexcept {
+	void Point::SetY(float y) noexcept {
 		m_y = y;
 	}
 
@@ -39,11 +39,11 @@ namespace Utility {
 		return Point(lhs.GetX() / rhs.GetX(), lhs.GetY() / rhs.GetY());
 	}
 
-	Point operator*(const Point& lhs, const int& rhs) {
+	Point operator*(const Point& lhs, const float& rhs) {
 		return Point(lhs.GetX() * rhs, lhs.GetY() * rhs);
 	}
 
-	Point operator/(const Point& lhs, const int& rhs) {
+	Point operator/(const Point& lhs, const float& rhs) {
 		return Point(lhs.GetX() / rhs, lhs.GetY() / rhs);
 	}
 }
